Sadly include all our puppet modules in the repo
I don't like doing this for a few reasons, but it's so much easier for
new folks to start out if they don't have to also install puppet on
their local machines in order to provision their VMs.

I am erring on the side of making this easier for folks to get up and
running with. The trade-off is that when I make provisioning changes,
the diffs will be larger, louder, and more annoying. After helping
several folks get this provisioning stuff working, I'm leaning more and
more on the side of 'just commit it.