Exception handling for IntegrityError
If a user tries to enter a tag that she has already entered,
she will violate the unique constraint - this is an IntegrityError.
If the (officer_id, img_id) link already exists, rollback and then
flash a message to the user indicating that a link already exists
for that officer so it was not entered again.