Separate migration management command into makemigrations and migrate
sqlalchemy-migrate does not always generate migrations that work.
Occasionally modifications to the generated migrations must be done.
I've separated the migration command into one that is run by developers
"makemigrations" that generates the migrations. These can then be
edited if necessary. Then "migrate" is executed in order to actually
apply these migrations to the database.