Add "Skip this image" buttons to the classification and identification tools
It's annoying if you get an image you don't really know what to do with
(if we get a better workflow for officers that we can't find in the officer
then hopefully people won't be using this much) that you can't skip
an image. This commit adds a skip button to `/cop_face` and `/sort` views.