Make the fabfile backup and deploy better
DEPLOY.md (and, coincidentally, OO in staging and prod) asserts the
existence of a virtualenv but doesn't actually manage it very well. So,
for instance, the `pip install` command in `fabfile.py` installs
`requirements.txt`, but for some reason has a good chance of doing it in
the system python context. Which is not ideal, to say the least, because
user `nginx` can't write to `/usr` to populate these, and people who
think they've deployed a new dep may see the deployed service fail to
start. This change makes the virtualenv usage a little more explicit:
we're spawning a gunicorn installed and managed by the virtualenv, and
we're running the pip/python explicitly from the venv.

Also, the backup dumped the sql twice and gave it a name that
didn't sort lexicographically. This fixes that.