Incident form: Show optional fields and make most fields optional
Minor tweak.

Note the fix for lack of time is a hack, I don't have time
to do a better job here. This codes the times such that we can
use this hardcoded time (which you'll note sets the time down to
the millisecond, something that is not possible via the UI so
we can cleanly select incidents that were entered without a time).