requirements: specify in requirements to install from source
Previously we were getting a warning about the fact that the built
wheel packages for psycopg2 are being distributed under another name,
psycopg2-binary. These wheels are recommended in the docs not for
production use, and additionally introduce some issues due to the
fact that many projects are depending on psycopg2, not psycopg2-binary.
Let's keep what we have now and suppress the warning via specifying
--no-binary in the requirements file.

See:
https://github.com/psycopg/psycopg2/issues/674#issuecomment-434352029
http://initd.org/psycopg/docs/install.html#binary-install-from-pypi
https://github.com/lucyparsons/OpenOversight/pull/618