Test python versions we think we're testing
There's a regression in 3.8.4, which broke all 3 of our builds despite
the travis build matrix. This is because we weren't actually running the
tests using those versions, because the tests run inside docker where
the version was specified by the Dockerfile to be 3.8.

Luckily, flake8 was still running in the travis python version, so we
caught things like f-strings not being supported in 3.5.

I expect this will fail on 3.8 and pass on 3.5, 3.6, and 3.7, but I won't know
until it's pushed. Once that's verified to work, we can drop 3.8 support
until we're running an OS somewhere that needs it or the image is fixed.