Test with the python versions we think we're testing (#757)
There's a regression in 3.8.4, which broke all 3 of our builds despite
the travis build matrix. This is because we weren't actually running the
tests using those versions, because the tests run inside docker where
the version was specified by the Dockerfile to be 3.8.

Luckily, flake8 was still running in the travis python version, so we
caught things like f-strings not being supported in 3.5.

I expect this would now fail on 3.8 and pass on 3.5, 3.6, and 3.7, but I was
not planning on adding 3.8 support to the matrix yet (and am less inclined
to do so while this bug is unfixed).

While I was in here I also cleaned up the Dockerfile a little bit.