Upgrade Flask-SQLAlchemy to guard against #754
In looking at flask-sqlalchemy's github issue about this, I realized
that:
a) they already pushed a fixed release
b) this will affect anyone who happens to be running a 3.8.4, which
shouldn't be too many people, but may be nonzero

Let's just add the fix to the package, just in case. And while I'm here,
let's test against 3.8 again too.