Upgrade to latest NodeJS LTS
Docker was complaining about the version of nodejs we were running just because it stopped being supported in December. Upgrade to latest nodejs 12 LTS.
Also, upgrade the version of node-sass we use to a version compatible with nodejs 12.

No visible differences in our generated CSS, though I did not do a before/after CSS diff.