Merge pull request from GHSA-gfjg-2w9g-vgj3
* Make a constant for the JWT signing algorithm

* add users._uuid column

This column will provide a unique identifier that can be used to
lookup users.

* Add uuid property to User

The _uuid column is prefixed with an underscore to deter developers from
modifying it outside of specific circumstances. Providing a decorator to
access the value of _uuid reads better, and further discourages
interacting with _uuid directly.

* Use uuid instead of id in JWT payloads

JWTs serve to authenticate requests that confirm or change a user's
email address, as well as requests to change a user's password.

In some circumstances, it would be good to invalidate a JWT. For actions
that use JWTs, the application validates the request by checking that
the JWT was signed by the correct key, that it has not expired, and that
the claims contained in the JWT are valid.

Rotating the signing key would invalidate all JWTs signed by that key,
which is undesirable. However, we can render specific JWTs invalid by
changing the value of the user ID contained in the payload claims. This,
too, is undesirable, because the user ID is referenced in database
associations. Since the user's UUID provides the same capability of
uniquely identifying a user, we can use it in place of the ID in the
JWT payload, which gives us an easily-rotated attribute that won't cause
complications when changed.

* Invalidate pw reset JWT after pw changed

This change ensures a password reset token cannot be reused after the
user has successfully changed their password.

* Ensure pw reset token invalidated on email change

This commit regenerates the user's UUID when their email address
changes, so that any JWTs sent to the old email address would no
longer be valid.

* Add comment explaining UUID concept

* Regenerate user UUID when setting password

In addition to regenerating the UUID when a password is reset via the
forgot password flow, this commit regenerates the UUID any time the
password is set. This ensures a password reset JWT is invalidated in
a scenario where a user has generated the password reset JWT, but then
remembers their password and resets it via the normal password change
flow.

Prior to this change, a new user would only have their UUID assigned
when the changes were committed to the database, but now the UUID gets
populated as part of the password hash assignment. A few tests have been
updated to reflect this new behavior.

---------

Co-authored-by: Michael Plunkett <5885605+michplunkett@users.noreply.github.com>