Timezone-related improvements (#1055)
<!-- New Contributor? Welcome!

We recommend you check your privacy settings, so the name and email
associated with
the commits are what you want them to be. See the contribution guide at

https://github.com/lucyparsons/OpenOversight/blob/develop/CONTRIB.md#recommended-privacy-settings
for more infos.

Also make sure you have read and abide by the code of conduct:

https://github.com/lucyparsons/OpenOversight/blob/develop/CODE_OF_CONDUCT.md

If this pull request is not ready for review yet, please submit it as a
draft.
-->
Several misc timezone-related changes:
* Validate input to /timezone
* Remove pytz dependency
https://blog.ganssle.io/articles/2018/03/pytz-fastest-footgun.html
* Include timezone in local_time and local_date_time output
* Add tests for jinja filters
* Fix typo `thousands_seperator` -> `thousands_separator`
* Revert change to make session cookie permanent

None!

N/A

 - [x] This branch is up-to-date with the `develop` branch.
 - [x] `pytest` passes on my local development environment.
 - [x] `pre-commit` passes on my local development environment.