Deploy to prod on tag creation (#1064)
## Fixes issue
#1030

## Description of Changes
Previously a push on the main branch triggered a deploy to prod. In the
future, creating a release will trigger that push.
The change is already built in a way that we can delete the main branch
and rename the develop branch to main as a next step (once we confirmed
everything works).