Remove broken DB details link (#1092)
## Fixes issue
https://github.com/lucyparsons/OpenOversight/issues/1091

## Description of Changes
Removed a link to a non-existent page. Since the schema underwent
significant changes over the last summer, it's likely that even if they
link was active it wouldn't be up date date. The current picture in the
`README` gives an in-depth look at the relationships between tables,
etc. What we have now should be sufficient.