Revert "Delete `create_db.py` (#1000)" (#1152)
## Fixes issue

#1000 unfortunately broke alembic's functionality to create migrations
that include new tables (Using `flask db migrate`). The issue is that
before alembic has a change to analyze the difference between existing
database and schema from database.py the `db.create_all()` command
automatically creates the tables making it so that they are not
registered for the migration.

I could not come up with a way to reasonably work around this issue, so
I think the best course of action is to revert the removal of the
create_db.py file.


## Description of Changes

This reverts that breaking commit. It is not a clean revert of #1000,
since that PR also included unrelated changes to tests and constants.
Those changes were not reverted.


## Notes for Deployment

I have verified the changes in my local environment and the tests are
passing, but I have not tested that the changes work as expected in a
completely fresh environment (which would use the create_db file)

## Tests and Linting
- [x] This branch is up-to-date with the `develop` branch.
- [ ] (N/A) Ran `make create_db_diagram` command.
- [x] `pytest` passes on my local development environment.
- [x] `pre-commit` passes on my local development environment.

---------

Co-authored-by: michplunkett <5885605+michplunkett@users.noreply.github.com>