Overhaul the `/departments/<department_id>` view (#1164)
## Description of Changes

I ran into this when working on #1148

The BrowseForm was basically only used as a data-container, this change
is making sure we also use it to filter the incoming data to simplify
the code in the view file. I also included other improvements like
reducing the number of SQL queries sent.

Functionality wasn't changed (at least that was the goal), so the user
experience should remain the same.