Remove `docker-compose` versions (#1178)
## Description of Changes
The docker-compose `version` tag is now obsolete, so I removed it from
the files.

> The top-level version property is defined by the Compose Specification
for backward compatibility. It is only informative you'll receive a
warning message that it is obsolete if used.

> Compose doesn't use version to select an exact schema to validate the
Compose file, but prefers the most recent schema when it's implemented.

Source:
[Link](https://github.com/compose-spec/compose-spec/blob/main/04-version-and-name.md)

## Tests and Linting
- [x] This branch is up-to-date with the `develop` branch.
- [x] `pytest` passes on my local development environment.
- [x] `pre-commit` passes on my local development environment.
- [x] `build_deploy` still works:
[Link](https://github.com/lucyparsons/OpenOversight/actions/runs/14455440793/job/40537505981?pr=1178)
- It failed because the there was no PR to merge, but it otherwise
seemed fine. Docker did not immediately error out.