Move `CODEOWNERS` file (#1184)
## Description of Changes
I'd like to remove the amount of clutter in the base directory, so I
moved the `CODEOWNERS` file.

> To use a CODEOWNERS file, create a new file called CODEOWNERS in the
.github/, root, or docs/ directory of the repository, in the branch
where you'd like to add the code owners. If CODEOWNERS files exist in
more than one of those locations, GitHub will search for them in that
order and use the first one it finds.

Source:
[link](https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/about-code-owners#codeowners-file-location)

## Tests and Linting
- [x] This branch is up-to-date with the `develop` branch.
- [x] `pytest` passes on my local development environment.
- [x] `pre-commit` passes on my local development environment.